setup docker and docker-compose also, change file structure, split to folders
