+++ practice-app/backend/requirements.txt
@@ -1,0 +1,3 @@
+Django>=3.0,<4.0
+psycopg2-binary>=2.8
+djangorestframework
