+++ practice-app/frontend/Dockerfile
@@ -1,0 +1,9 @@
+FROM node:alpine
+
+WORKDIR /app/frontend
+COPY . /app/frontend
+
+RUN npm install
+
+EXPOSE 3000
+# CMD ["npm", "start"]
