+++ practice-app/frontend/public/manifest.json
